Display raid duration when there are two timers
